move _get_sample_intervals to pvlib.tools
